requirements.txt
transformers>=4.15.0
requests>=2.27.1
Pillow~=9.0.0
torchvision~=0.11.2
numpy>=1.22.1
torch>=1.10.1
imageio
